Cleaned up existing test cases - Refactored and created file for mocks
